feat: structure change more like a package style
