+++ noqx.py
@@ -1,45 +1,11 @@
 """Main file for the noqx project."""
 
 import argparse
-import traceback
-from typing import Any, Dict
 
 import uvicorn
-from starlette.applications import Starlette
-from starlette.requests import Request
-from starlette.responses import JSONResponse, RedirectResponse
-from starlette.routing import Mount, Route
-from starlette.staticfiles import StaticFiles
+from starlette.config import environ
 
-from solver import list_solver_metadata, run_solver
-from solver.core.logging import logger
 from solver.core.solution import Config
-
-
-async def root_redirect(_: Request) -> RedirectResponse:
-    """Redirect root page to penpa-edit page."""
-    return RedirectResponse(url="/penpa-edit/")
-
-
-async def list_puzzles_api(_: Request) -> JSONResponse:
-    """List the available puzzles."""
-    return JSONResponse(list_solver_metadata())
-
-
-async def solver_api(request: Request) -> JSONResponse:
-    """The solver endpoint of the server."""
-    try:
-        body = await request.json()
-        puzzle_type: str = body["puzzle_type"]
-        puzzle: str = body["puzzle"]
-        param: Dict[str, Any] = body["param"]
-        return JSONResponse(run_solver(puzzle_type, puzzle, param))
-    except AssertionError as err:
-        logger.error(traceback.format_exc())
-        return JSONResponse({"detail": str(err)}, status_code=400)
-    except TimeoutError as err:
-        return JSONResponse({"detail": str(err)}, status_code=504)
-
 
 parser = argparse.ArgumentParser(description="noqx startup settings.")
 parser.add_argument("-H", "--host", default="127.0.0.1", type=str, help="the host to run the server on.")
@@ -51,20 +17,6 @@ args = parser.parse_args()
 Config.time_limit = args.time_limit
 Config.parallel_threads = args.parallel_threads
 
-routes = [
-    Mount(
-        "/api",
-        name="api",
-        routes=[
-            Route("/list/", endpoint=list_puzzles_api, methods=["GET"]),
-            Route("/solve/", endpoint=solver_api, methods=["POST"]),
-        ],
-    ),
-    Mount("/penpa-edit/", StaticFiles(directory="static", html=True), name="static"),
-    Route("/", root_redirect),
-]
-
-app = Starlette(debug=args.debug, routes=routes)
-
 if __name__ == "__main__":
+    environ["DEBUG"] = "TRUE" if args.debug else "FALSE"
     uvicorn.run(app="noqx:app", host=args.host, port=args.port, reload=args.debug, log_level="debug" if args.debug else "info")
